SM-1268: Improving Java lang example and docs
