+++ languages/java/settings.gradle
@@ -3,3 +3,5 @@
  */
 
 rootProject.name = 'BitwardenSDK'
+
+include "example"
